+++ src/main/java/com/example/ecommerce/DTO/response/OrderItemResponseDTO.java
@@ -8,11 +8,11 @@ import java.time.LocalDateTime;
 @Builder
 public class OrderItemResponseDTO {
     private String name;
-    private Double price;
-    private Double discountPrice;
+    private Long price;
+    private Long discountPrice;
     private String imageUrl;
     private Integer quantity;
-    private Double totalPrice;
+    private Long totalPrice;
     private LocalDateTime createdOn;
     private LocalDateTime updatedOn;
 }
